feat: refactor auth and cache directory
