+++ cache/instance_test.go
