+++ service/service.go
@@ -21,8 +21,8 @@ import (
 	"context"
 	"fmt"
 	"github.com/golang/protobuf/ptypes/wrappers"
+	"github.com/polarismesh/polaris-server/cache"
 	commontime "github.com/polarismesh/polaris-server/common/time"
-	"github.com/polarismesh/polaris-server/service/cache"
 	"github.com/polarismesh/polaris-server/store"
 	"time"
